pyproject.toml
[project]
name = "babulletinbotv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
]
